Move away from the SiteDomain type
