+++ client/dashboard/emails/fields/domain-name.ts
@@ -1,9 +1,9 @@
-import { SiteDomain } from '@automattic/api-core';
+import { Domain } from '@automattic/api-core';
 import { __ } from '@wordpress/i18n';
 import type { Email } from '../types';
 import type { Field } from '@wordpress/dataviews';
 
-export const getDomainNameField = ( domains: SiteDomain[] ): Field< Email > => ( {
+export const getDomainNameField = ( domains: Domain[] ): Field< Email > => ( {
 	id: 'domainName',
 	label: __( 'Domain' ),
 	getValue: ( { item }: { item: Email } ) => item.domainName,
