+++ client/dashboard/sites/domains/primary-domain-selector.tsx
@@ -1,4 +1,4 @@
-import { type SiteDomain, type Site, type User, DomainSubtype } from '@automattic/api-core';
+import { type DomainSummary, type Site, type User, DomainSubtype } from '@automattic/api-core';
 import { siteSetPrimaryDomainMutation } from '@automattic/api-queries';
 import { useMutation } from '@tanstack/react-query';
 import {
@@ -19,7 +19,7 @@ import { userHasFlag } from '../../utils/user';
 import type { Field } from '@wordpress/dataviews';
 
 interface PrimaryDomainSelectorProps {
-	domains: SiteDomain[];
+	domains: DomainSummary[];
 	site: Site;
 	user: User;
 }
@@ -41,23 +41,17 @@ const PrimaryDomainSelector = ( { domains, site, user }: PrimaryDomainSelectorPr
 		if ( ! domains || ! site ) {
 			return [];
 		}
-		const hasWpcomStagingDomain = domains.find( ( domain ) => domain.is_wpcom_staging_domain );
 		return domains.filter( ( domain ) => {
 			// Basic eligibility criteria
 			const isEligible =
 				( domain.subtype.id === DomainSubtype.DOMAIN_REGISTRATION ||
-					domain.subtype.id === DomainSubtype.DOMAIN_CONNECTION ) &&
+					domain.subtype.id === DomainSubtype.DOMAIN_CONNECTION ||
+					domain.subtype.id === DomainSubtype.DEFAULT_ADDRESS ) &&
 				domain.can_set_as_primary &&
-				! domain.primary_domain &&
-				! domain.aftermarket_auction;
+				! domain.primary_domain;
 
 			if ( ! isEligible ) {
 				return false;
-			}
-
-			// Additional filtering for wpcom staging domains
-			if ( hasWpcomStagingDomain && domain.wpcom_domain ) {
-				return domain.is_wpcom_staging_domain;
 			}
 
 			return true;
