+++ client/dashboard/utils/domain-types.ts
@@ -1,6 +1,6 @@
 import { DomainSubtype } from '@automattic/api-core';
-import type { SiteDomain } from '@automattic/api-core';
+import type { DomainSummary } from '@automattic/api-core';
 
-export function isTransferrableToWpcom( domain: SiteDomain ) {
+export function isTransferrableToWpcom( domain: DomainSummary ) {
 	return domain.subtype.id === DomainSubtype.DOMAIN_CONNECTION;
 }
