+++ client/tsconfig.json
@@ -1,6 +1,7 @@
 {
 	"extends": "@automattic/calypso-typescript-config/mixed-package.json",
 	"compilerOptions": {
+		"incremental": true,
 		"rootDir": ".",
 		"noEmit": true,
 		"types": [
