+++ packages/api-core/src/site-domains/mutators.ts
@@ -1,6 +1,5 @@
 import { wpcom } from '../wpcom-fetcher';
-import type { SiteDomain } from './types';
 
-export async function setPrimaryDomain( siteId: number, domain: string ): Promise< SiteDomain > {
-	return wpcom.req.post( `/sites/${ siteId }/domains/primary`, { domain } );
+export async function setPrimaryDomain( siteId: number, domain: string ): Promise< void > {
+	await wpcom.req.post( `/sites/${ siteId }/domains/primary`, { domain } );
 }
